+++ channels/rail/client/CMakeLists.txt
@@ -18,6 +18,8 @@
 define_channel_client("rail")
 
 set(${MODULE_PREFIX}_SRCS
+    ../rail_common.h
+    ../rail_common.c
 	rail_main.c
 	rail_main.h
 	rail_orders.c
