+++ channels/rail/client/rail_main.h
@@ -29,27 +29,7 @@
 
 #include <winpr/stream.h>
 
-struct rdp_rail_order
-{
-	rdpSettings* settings;
-	void* plugin;
-	RAIL_HANDSHAKE_ORDER handshake;
-	RAIL_CLIENT_STATUS_ORDER client_status;
-	RAIL_EXEC_ORDER exec;
-	RAIL_EXEC_RESULT_ORDER exec_result;
-	RAIL_SYSPARAM_ORDER sysparam;
-	RAIL_ACTIVATE_ORDER activate;
-	RAIL_SYSMENU_ORDER sysmenu;
-	RAIL_SYSCOMMAND_ORDER syscommand;
-	RAIL_NOTIFY_EVENT_ORDER notify_event;
-	RAIL_MINMAXINFO_ORDER minmaxinfo;
-	RAIL_LOCALMOVESIZE_ORDER localmovesize;
-	RAIL_WINDOW_MOVE_ORDER window_move;
-	RAIL_LANGBAR_INFO_ORDER langbar_info;
-	RAIL_GET_APPID_REQ_ORDER get_appid_req;
-	RAIL_GET_APPID_RESP_ORDER get_appid_resp;
-};
-typedef struct rdp_rail_order rdpRailOrder;
+#include "../rail_common.h"
 
 struct rail_plugin
 {
@@ -61,10 +41,6 @@ typedef struct rail_plugin railPlugin;
 void rail_send_channel_event(void* rail_object, UINT16 event_type, void* param);
 void rail_send_channel_data(void* rail_object, void* data, size_t length);
 
-#ifdef WITH_DEBUG_RAIL
-#define DEBUG_RAIL(fmt, ...) DEBUG_CLASS(RAIL, fmt, ## __VA_ARGS__)
-#else
-#define DEBUG_RAIL(fmt, ...) DEBUG_NULL(fmt, ## __VA_ARGS__)
-#endif
+
 
 #endif /* FREERDP_CHANNEL_CLIENT_RAIL_MAIN_H */
